moved test server to another file
