+++ __tests__/index.test.js
@@ -1,46 +1,6 @@
 const MyAdminAPI = require('../index')
-const rest = require('msw').rest
-const setupServer = require('msw/node').setupServer
+const server = require('../server-for-tests')
 
-const server = setupServer(
-	rest.post('https://myadminapitest.geotab.com/v2/MyAdminApi.ashx', (req, res, ctx) => {
-		switch(req.body.method) {
-			case 'GetCountries': {
-				return res(
-					ctx.status(200),
-					ctx.json({ result: ['Canada', 'Philippines', 'United States'] })
-					)
-				}
-			break
-			case '': {
-				return res(
-				ctx.status(500),
-				ctx.json({
-					result: {
-						error: {
-							message: 'An error has occurred and this string contains info.',
-							name: 'MissingMethodException'
-						}
-					}
-				})
-				)
-			}
-			break
-			default: {
-				return res(
-					ctx.status(200), 
-					ctx.json({
-						result: {
-							among_other_things: 'this result contains sessionId and userId, below',
-							sessionId: 'a session Id',
-							userId: 'a userId',
-						}
-					})
-				)		
-			}
-		}
-	})
-)
 
 beforeAll(() => {
 	server.listen()
